2024-09-10 11:50 PM
[Chat]
{ChatRoomController}
::getUserInfo를 Jwt 용 UserInfoController로 이동
{ChatRoomRedisRepository}
설명 주석 추가 및 enterChatRoom과 getTopic 삭제
{ChatRoomRepository}
findByTeam 메서드 변경
{ChatRoomService}
::enterChatRoom 제거
{StompHandler}
{JWT}
{UserInfoController}
::getUserInfo

[JS]
roomdetail
Acked-by: 구본우
