+++ src/main/resources/application.yml
@@ -14,6 +14,8 @@ spring:
   redis:
     host: localhost
     port: 6379
+  jwt:
+    secret: specifiedsecretchat!@#
 
   #STATIC 파일 자동 반영
   devtools:
